- data-collector   all non-coremedia services (mysql, node_exporter, etc.) can now be in the cm-service.yaml configured   e.g. port and cretentials
  ```
  mysql:
    description: MySQL
    port: 3306
    monitoring_user: monitoring
    monitoring_password: monitoring
  ```

- service-discovery
  first implement of a 'rediscovery' feature

- docker-compose
  remove all hard coded network ip
  use now the DNS from docker-host
  put all environment variables from environment.yamlbackto docker-compose.yaml (possible depricated)

- grafana-client
  update to grafana gem 0.6.6

+++ osmc/blueprint-box.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-set -e
+# set -e
 
 osmc_deploy() {
 
@@ -100,9 +100,9 @@ osmc_deploy() {
 
   [ -f /tmp/deployment-archive.zip ] && cp -a /tmp/deployment-archive.zip /var/tmp/deploy/deployment-archive.zip
 
-  curl -L -o cms.zip  "https://repository-build.coremedia.com/nexus/service/local/artifact/maven/redirect?g=com.coremedia.cms.license&a=development-license&v=17-SNAPSHOT&p=zip&r=snapshots.licenses&c=prod"
-  curl -L -o mls.zip  "https://repository-build.coremedia.com/nexus/service/local/artifact/maven/redirect?g=com.coremedia.cms.license&a=development-license&v=17-SNAPSHOT&p=zip&r=snapshots.licenses&c=pub"
-  curl -L -o rls.zip  "https://repository-build.coremedia.com/nexus/service/local/artifact/maven/redirect?g=com.coremedia.cms.license&a=development-license&v=17-SNAPSHOT&p=zip&r=snapshots.licenses&c=repl"
+  [ -f cms.zip ] || curl -L -o cms.zip  "https://repository-build.coremedia.com/nexus/service/local/artifact/maven/redirect?g=com.coremedia.cms.license&a=development-license&v=17-SNAPSHOT&p=zip&r=snapshots.licenses&c=prod"
+  [ -f mls.zip ] || curl -L -o mls.zip  "https://repository-build.coremedia.com/nexus/service/local/artifact/maven/redirect?g=com.coremedia.cms.license&a=development-license&v=17-SNAPSHOT&p=zip&r=snapshots.licenses&c=pub"
+  [ -f rls.zip ] || curl -L -o rls.zip  "https://repository-build.coremedia.com/nexus/service/local/artifact/maven/redirect?g=com.coremedia.cms.license&a=development-license&v=17-SNAPSHOT&p=zip&r=snapshots.licenses&c=repl"
 
   # jetzt gehts loooooo ... hos
   [ -f deployment-archive.zip ] && unzip -o deployment-archive.zip
@@ -122,6 +122,7 @@ osmc_deploy() {
   sleep 4s
 
   # enable mod_status
+  echo "Listen *:8081" >> /etc/httpd/ports.conf
   sed -i 's|ExtendedStatus Off|ExtendedStatus On|g' /etc/httpd/mods-available/status.conf
   sed -i 's|Require local|Require all granted|g' /etc/httpd/sites-available/_overview.conf
 
@@ -197,14 +198,7 @@ after_deploy() {
   #done
 }
 
-reset_dard() {
-
-  cd /opt/coremedia/caefeeder-preview-tools/bin/
-  ./cm resetcaefeeder reset
-
-  cd /opt/coremedia/caefeeder-live-tools/bin/
-  ./cm resetcaefeeder reset
-
+reset_hard() {
 
   for i in caefeeder-live caefeeder-preview cae-live-1 cae-preview content-feeder content-management-server mongod editor-webapp elastic-worker master-live-server replication-live-server sitemanager solr studio user-changes workflow-server mysql-default
   do
@@ -213,25 +207,56 @@ reset_dard() {
 
   find /var/log/coremedia/ -type f -exec rm -f {} \;
 
-  yum erase mysql-community-common mysql-community-client mysql-community-libs-compat mysql-community-libs mysql-community-server mysql-community-devel
+  yum -y erase mysql-community-common mysql-community-client mysql-community-libs-compat mysql-community-libs mysql-community-server mysql-community-devel
 
   rm -rf /var/lib/mysql-default
   rm -rf /var/tmp/coremedia
+  rm -rf /var/lib/mongo/*
+  rm -rf /var/coremedia/solr-data
 
-  exec /var/tmp/deploy/osmc-deploy.sh
+  cd /var/tmp/deploy
 
+  [ -f deployment-archive.zip ] && unzip -o deployment-archive.zip
+
+  service mongod start
+
+  /var/tmp/deploy/osmc-deploy.sh
+
+  # TODO
+  # sed -i 's|#SOLR_HOST="192.168.1.1"|SOLR_HOST=\"${HOSTNAME}\"|g' /etc/default/solr.in.sh
+  sed -i 's|ENABLE_REMOTE_JMX_OPTS="false"|ENABLE_REMOTE_JMX_OPTS="true"|g' /opt/solr/bin/solr.in.sh
+
+  service solr restart
+
+  # enable mod_status
+  echo "Listen *:8081" >> /etc/httpd/ports.conf
+  sed -i 's|ExtendedStatus Off|ExtendedStatus On|g' /etc/httpd/mods-available/status.conf
+  sed -i 's|Require local|Require all granted|g' /etc/httpd/sites-available/_overview.conf
+
+  /opt/chef/embedded/bin/ruby /tmp/apache_vhosts.rb > /opt/coremedia/overview/vhosts.json
+
+  service httpd reload
 
   sleep 10s
 
+#  /opt/coremedia/content-management-server-tools/bin/cm publishall -a -cq "NOT BELOW PATH '/Home'" -t 1 http://${HOSTNAME}:40180/coremedia/ior admin admin http://${HOSTNAME}:40280/coremedia/ior admin admin
+#  /opt/coremedia/content-management-server-tools/bin/cm serverimport -r -u admin -p admin --no-validate-xml -t 4 /var/tmp/coremedia/test-data/content
+#  sleep 4s
 
-  cd /opt/coremedia/replication-live-server-tools/bin/
-  ./cm runlevel -u admin -p admin
-  ./cm runlevel -u admin -p admin -r online -g 2
+  /opt/coremedia/content-management-server-tools/bin/cm runlevel -u admin -p admin
+  /opt/coremedia/master-live-server-tools/bin/cm runlevel -u admin -p admin
+  /opt/coremedia/replication-live-server-tools/bin/cm runlevel -u admin -p admin
+  /opt/coremedia/replication-live-server-tools/bin/cm runlevel -u admin -p admin -r online -g 2
+  /opt/coremedia/caefeeder-preview-tools/bin/cm resetcaefeeder reset
+  /opt/coremedia/caefeeder-live-tools/bin/cm resetcaefeeder reset
 
+  service caefeeder-live restart
+  service caefeeder-preview restart
 
+  tail -F /var/log/coremedia/*/*log
 }
 
 
-osmc_deploy
+reset_hard
 
 # shutdown -rF now
